requirements.txt
argparse
requests
tqdm
beautifulsoup4
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
colorama